Business, People & Employment Data/ACQ Members Scraper/acq_scraper/requirements.txt
itemadapter==0.8.0
pandas==2.2.2
python-slugify==8.0.4
Scrapy==2.9.0
